PASSWORD GENERATOR
The password generator project in Python involves creating a tool that takes user input for the desired password length. It then uses a combination of random characters to generate a strong and secure password of the specified length. The generated password is displayed on the screen for the user. This project is helpful for creating robust passwords with varying lengths and complexities, enhancing user security in online accounts and systems.